% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{cell_type_marker_labels}
\alias{cell_type_marker_labels}
\title{Cell type marker labels
Display labels for \code{cell_type_marker_genes}.}
\format{
A named character vector.
}
\usage{
cell_type_marker_labels
}
\description{
Cell type marker labels
Display labels for \code{cell_type_marker_genes}.
}
\keyword{datasets}
