% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{CONTRAST_DISPLAY_LABEL}
\alias{CONTRAST_DISPLAY_LABEL}
\title{Parenthetical contrast label for plot axes/titles.}
\usage{
CONTRAST_DISPLAY_LABEL
}
\description{
Parenthetical contrast label for plot axes/titles.
}
