% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{palette_analysis_three}
\alias{palette_analysis_three}
\title{Three-color analysis palette for depleted/neutral/enriched (low/mid/high) encodings.}
\usage{
palette_analysis_three
}
\description{
Three-color analysis palette for depleted/neutral/enriched (low/mid/high) encodings.
}
