% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{palette_dotplot_pair}
\alias{palette_dotplot_pair}
\title{Two-color low/high analysis palette; subset of \code{palette_analysis_three}.}
\usage{
palette_dotplot_pair
}
\description{
Two-color low/high analysis palette; subset of \code{palette_analysis_three}.
}
