% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-abundance-plots.R
\name{plot_clr_fisher_enrichment}
\alias{plot_clr_fisher_enrichment}
\title{Plot CLR/Fisher cluster enrichment.}
\usage{
plot_clr_fisher_enrichment(enrichment)
}
\arguments{
\item{enrichment}{Data frame returned by \code{compute_cluster_abundance()}.}
}
\value{
A ggplot object.
}
\description{
Plot CLR/Fisher cluster enrichment.
}
