% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-abundance.R
\name{plot_cluster_proportion_by_mouse}
\alias{plot_cluster_proportion_by_mouse}
\title{Plot sample-level cluster proportions by mouse.}
\usage{
plot_cluster_proportion_by_mouse(
  sample_props,
  control_label = CTRL_LABEL,
  estim_label = ESTIM_LABEL
)
}
\arguments{
\item{sample_props}{Data frame returned by
\code{compute_sample_cluster_proportions()}.}

\item{control_label}{Control condition label.}

\item{estim_label}{E-Stim condition label.}
}
\value{
A ggplot object.
}
\description{
Plot sample-level cluster proportions by mouse.
}
