% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{theme_stone}
\alias{theme_stone}
\title{Theme for publication figures}
\usage{
theme_stone(base_size = 12)
}
\arguments{
\item{base_size}{Base font size.}
}
\value{
A ggplot2 theme.
}
\description{
Theme for publication figures
}
