% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{cell_type_marker_genes}
\alias{cell_type_marker_genes}
\title{Cell type marker genes}
\format{
A named list of mouse gene symbols.
}
\usage{
cell_type_marker_genes
}
\description{
Curated marker genes for broad retinal cell type annotation.
}
\keyword{datasets}
