% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{mouse_cell_cycle_genes}
\alias{mouse_cell_cycle_genes}
\title{Mouse cell-cycle genes}
\format{
A character vector of mouse gene symbols.
}
\usage{
mouse_cell_cycle_genes
}
\description{
Mouse orthologs of Seurat's human cell-cycle genes, generated with
\code{data-raw/mouse-cell-cycle-genes.R} using \code{biomaRt}.
}
\keyword{datasets}
